+++ src/main/java/me/flotsam/frettler/command/BassGuitarCommand.java
@@ -30,6 +30,6 @@ public class BassGuitarCommand extends FrettedInstrumentCommand implements Runna
 
   @Override
   public void run() {
-   exec(new BassGuitar(strings, frets));
+   exec(new BassGuitar(strings, tuning, frets));
   }
 }