+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2nd May 2021
+- Added out of the box `--tuning` support
+
+# 28th April 2021
+- Added left handed instrument support
+
 # 3rd February 2021
 - Added the progression subcommand